+++ web/lib/passport/index.ts
@@ -6,3 +6,4 @@ export * from './types';
 export * from './mrz';
 export * from './dates';
 export * from './persist';
+export * from './validity';